// src/utils/enums/providers.ts
export enum UOWProvider {
  UOW = 'IUnitOfWork',
}

export enum CustomerProvider {
  SERVICE = '',
  REPOSITORY = 'ICustomerRepository',
}

export enum PartnerProvider {
  SERVICE = '',
  REPOSITORY = 'IPartnerRepository',
}

export enum EventProvider {
  SERVICE = '',
  REPOSITORY = 'IEventRepository',
}

export enum OrderProvider {
  SERVICE = '',
  REPOSITORY = 'IOrderRepository',
}

export enum SpotReservationProvider {
  SERVICE = '',
  REPOSITORY = 'ISpotReservationRepository',
}